backend/users/requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
alembic==1.11.0
anyio==3.6.2
asyncclick==8.1.3.4
asyncpg==0.27.0
bcrypt==4.0.1
black==22.12.0
boto3==1.26.142
botocore==1.29.142
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
databases==0.7.0
dictdiffer==0.9.0
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.5.1
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
PyJWT==2.7.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.26.1
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.12.0
tortoise-orm==0.19.3
typing_extensions==4.5.0
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
